backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow>=10.2.0
qrcode[pil]>=7.4.2

# Optional database drivers (only install if needed)
# For PostgreSQL: pip install -r requirements-postgresql.txt
# For MySQL: pip install -r requirements-mysql.txt
# SQLite is included by default with Python, no driver needed

# Optional payment gateway SDKs (only install if needed)
# For Stripe: pip install stripe
# For PayPal: pip install paypalrestsdk
# These are optional - the payment_gateways.py module will work without them

